캘린더

+++ src/main/java/com/codenal/employee/repository/EmployeeRepository.java
@@ -1,13 +1,9 @@
 package com.codenal.employee.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
-
-import com.codenal.calendar.domain.Calendar;
 import com.codenal.employee.domain.Employee;
 
 public interface EmployeeRepository extends JpaRepository<Employee, Long> {
-    Employee findByEmpId(Long empId);
-    
+    // 필요시 커스텀 쿼리 추가 가능
+    Employee findByEmpId(int empId);
 }
